refine number classes
- no longer snupport gmpxx; it is either flint or imath
- refactor RationalQ class
- Integer and Rational classes still have to be cleaned up
