+++ CMakeLists.txt
@@ -23,18 +23,12 @@ include(CMakeDependentOption)
 
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 
-set(DEFAULT_CLINGOLPX_MATH_LIBRARY gmpxx)
-if (WIN32)
-    set(DEFAULT_CLINGOLPX_MATH_LIBRARY imath)
-endif()
-
 option(CLINGOLPX_MANAGE_RPATH "set rpath if not installed into system directory" ON)
 option(CLINGOLPX_BUILD_TESTS "build tests" OFF)
 option(CLINGOLPX_BUILD_SHARED "build clingo-lpx library shared" ON)
 option(CLINGOLPX_CROSSCHECK "enable additional state checking" OFF)
 option(CLINGOLPX_PROFILE "Enable profiling using gperftools" OFF)
-set(CLINGOLPX_MATH_LIBRARY "${DEFAULT_CLINGOLPX_MATH_LIBRARY}" CACHE STRING "math library: gmpxx, imath, flint")
-option(CLINGOLPX_USE_FLINT "use flint instead of gmpxx" ${WIN32})
+set(CLINGOLPX_MATH_LIBRARY "auto" CACHE STRING "math library: auto, imath, flint")
 option(CLINGOLPX_USE_LOCAL_CATCH
 "Advanced option to build against bundled or installed catch."
 ON)
@@ -101,18 +95,30 @@ endif()
 
 add_subdirectory(third_party)
 
-add_library(MATH::MATH INTERFACE IMPORTED)
-if (CLINGOLPX_MATH_LIBRARY STREQUAL "flint")
+if (CLINGOLPX_MATH_LIBRARY STREQUAL "auto")
+    find_package(FLINT)
+    if (FLINT_FOUND)
+        set(_CLINGOLPX_MATH_LIBRARY "flint")
+    else()
+        set(_CLINGOLPX_MATH_LIBRARY "imath")
+    endif()
+elseif (CLINGOLPX_MATH_LIBRARY STREQUAL "flint")
     find_package(FLINT REQUIRED)
+    set(_CLINGOLPX_MATH_LIBRARY "flint")
+elseif (CLINGOLPX_MATH_LIBRARY STREQUAL "imath")
+    set(_CLINGOLPX_MATH_LIBRARY "imath")
+else()
+    message(FATAL_ERROR "error unknown math library")
+endif()
+
+add_library(MATH::MATH INTERFACE IMPORTED)
+if (_CLINGOLPX_MATH_LIBRARY STREQUAL "flint")
     set_property(TARGET MATH::MATH PROPERTY INTERFACE_COMPILE_DEFINITIONS CLINGOLPX_USE_FLINT)
     set_property(TARGET MATH::MATH PROPERTY INTERFACE_LINK_LIBRARIES FLINT::FLINT)
-elseif (CLINGOLPX_MATH_LIBRARY STREQUAL "imath")
+elseif (_CLINGOLPX_MATH_LIBRARY STREQUAL "imath")
     set_property(TARGET MATH::MATH PROPERTY INTERFACE_COMPILE_DEFINITIONS CLINGOLPX_USE_IMATH)
     set_property(TARGET MATH::MATH PROPERTY INTERFACE_LINK_LIBRARIES libimath)
 else()
-    find_package(GMPXX REQUIRED)
-    set_property(TARGET MATH::MATH PROPERTY INTERFACE_COMPILE_DEFINITIONS CLINGOLPX_USE_GMPXX)
-    set_property(TARGET MATH::MATH PROPERTY INTERFACE_LINK_LIBRARIES GMPXX::GMPXX)
 endif()
 
 if (CLINGOLPX_PROFILE)
