+++ Makefile
@@ -22,7 +22,7 @@ compdb: configure
 configure: build/$(BUILD_TYPE)/build.ninja
 
 build/$(BUILD_TYPE)/build.ninja:
-	cmake -G Ninja -H. -B"build/$(BUILD_TYPE)" -DCMAKE_INSTALL_PREFIX=${HOME}/.local/opt/potassco/$(BUILD_TYPE) -DCMAKE_CXX_FLAGS="$(CXXFLAGS)" -DCMAKE_BUILD_TYPE="$(BUILD_TYPE)" -DCMAKE_CXX_CLANG_TIDY:STRING="$(CLANG_TIDY)" -DClingo_DIR="$(CLINGO_DIR)" -DCLINGOLPX_BUILD_TESTS=On -DCLINGOLPX_USE_FLINT=on -DCMAKE_EXPORT_COMPILE_COMMANDS=On
+	cmake -G Ninja -H. -B"build/$(BUILD_TYPE)" -DCMAKE_INSTALL_PREFIX=${HOME}/.local/opt/potassco/$(BUILD_TYPE) -DCMAKE_CXX_FLAGS="$(CXXFLAGS)" -DCMAKE_BUILD_TYPE="$(BUILD_TYPE)" -DCMAKE_CXX_CLANG_TIDY:STRING="$(CLANG_TIDY)" -DClingo_DIR="$(CLINGO_DIR)" -DCLINGOLPX_BUILD_TESTS=On -DCMAKE_EXPORT_COMPILE_COMMANDS=On
 
 Makefile:
 	:
